add employee report

+++ src/main/webapp/static/js/useryewu/employee_controller.js
@@ -1,21 +1,23 @@
 'use strict';
 
 
-app.controller('UserEmployeeController', ['$scope', 'EmployeeService', 'LoginService', 'usSpinnerService', '$location', 'MessageService', '$timeout', function ($scope, EmployeeService, LoginService, usSpinnerService, $location, MessageService, $timeout) {
+app.controller('UserEmployeeController', ['$scope', 'EmployeeService', 'LoginService', 'MessageService', '$timeout', 'UserFolderService', function ($scope, EmployeeService, LoginService, MessageService, $timeout, UserFolderService) {
     $scope.employees = [];
     $scope.selectedAll = false;
     $scope.itemPerPage = 5;
     $scope.editing = false;
-    $scope.editedEmployee={uuid:-1};
+    $scope.editedEmployee = {uuid: -1};
     var userId = LoginService.getUserInfo().userId;
     var newEmployee = {isNew: true};
     $scope.loadAll = function () {
-        EmployeeService.getEmployeeByUserId(LoginService.getUserInfo().userId).then(
+        $scope.employeeAllPromise = EmployeeService.getEmployeeByUserId(LoginService.getUserInfo().userId).then(
             function (data) {
                 var list = MessageService.handleMsg(data);
                 if (list) {
                     $scope.employees = list;
                     $scope.employees.push(newEmployee);
+                    getCustomizedField();
+
                 }
             }, function (err) {
                 MessageService.handleServerErr(err);
@@ -44,47 +46,41 @@ app.controller('UserEmployeeController', ['$scope', 'EmployeeService', 'LoginSer
     $scope.deleteEmployee = function (employee) {
         var ids = [];
         ids.push(employee.uuid);
-        EmployeeService.deleteEmployees(ids).then(function (response) {
-            var list = MessageService.handleMsg(response);
-            if (list) {
-                $scope.employees = list;
-            }
-        }, function (err) {
-            MessageService.handleServerErr(err);
-        });
+        var deleted = confirm("delete employee?");
+        if (deleted) {
+            EmployeeService.deleteEmployees(ids).then(function (response) {
+                var list = MessageService.handleMsg(response);
+                if (list) {
+                    $scope.employees = list;
+                }
+            }, function (err) {
+                MessageService.handleServerErr(err);
+            });
+        }
     };
 
-    //$scope.deleteSelectedEmployees = function () {
-    //    var ids = [];
-    //    angular.forEach($scope.employees, function (itm) {
-    //        if (itm.selected === true) {
-    //            ids.push(itm.uuid);
-    //        }
-    //    });
-    //    EmployeeService.deleteEmployees(ids).then(function (response) {
-    //        $scope.stopSpin();
-    //        $scope.loadAll();
-    //        $scope.employeeError.success = true;
-    //        $scope.employeeError.msg = "Success!";
-    //        $scope.employeeError.hasError = false;
-    //    }, function (err) {
-    //        $scope.stopSpin();
-    //        $scope.employeeError.hasError = true;
-    //        $scope.employeeError.success = false;
-    //        $scope.employeeError.msg = "Error while loading data, please try later!";
-    //    });
-    //};
     $scope.editEmployee = function (index) {
         $scope.editing = true;
-        $scope.editedEmployee = $scope.employees[index];
         $timeout(function () {
             var all = angular.element('#card-parent').children('div');
-            //angular.element(angular.element('#card-parent')).scrollTop(all[index].offsetTop);
-            $timeout(function () {
-                var all = angular.element('#card-parent').children('div');
-                angular.element('#card-parent').animate({'scrollTop':all[index].offsetTop},500);
-            });
+            angular.element('#card-parent').animate({'scrollTop': all[index].offsetTop}, 500);
         });
+        $scope.employeeEditPromise = EmployeeService.getEmployee(LoginService.getUserInfo().userId, $scope.employees[index].uuid)
+            .then(
+            function (msg) {
+                var employee = MessageService.handleMsg(msg);
+                if (employee) {
+                    $scope.editedEmployee = employee;
+                    setCustomizedEmployeeFieldsValue($scope.customizedEmployeeFields, employee.records);
+                    //if($scope.viewingReports){
+                    //    $scope.viewReports($scope.editedEmployee);
+                    //}
+                }
+            },
+            function (err) {
+                MessageService.handleServerErr(err);
+            }
+        );
     };
 
     $scope.saveEmployee = function (employee) {
@@ -96,13 +92,19 @@ app.controller('UserEmployeeController', ['$scope', 'EmployeeService', 'LoginSer
     };
 
     var createEmployee = function (employee) {
-        EmployeeService.createEmployee(userId, employee).then(
+        var records = [];
+        for (var i = 0; i < $scope.customizedEmployeeFields.length; i++) {
+            var record = new Record($scope.customizedEmployeeFields[i].id, employee.uuid, LoginService.getUserInfo().userId, $scope.customizedEmployeeFields[i].value);
+            records.push(record);
+        }
+        employee.records = records;
+        $scope.employeeEditPromise = EmployeeService.createEmployee(userId, employee).then(
             function (data) {
                 var list = MessageService.handleMsg(data);
                 if (list) {
-                    $scope.editedEmployee={};
+                    $scope.editedEmployee = {uuid: -1};
                     $scope.employees = list;
-                    newEmployee = {isNew:true};
+                    newEmployee = {isNew: true};
                     $scope.employees.push(newEmployee);
                     $scope.editing = false;
                 }
@@ -114,13 +116,19 @@ app.controller('UserEmployeeController', ['$scope', 'EmployeeService', 'LoginSer
     };
 
     var updateEmployee = function (employee) {
-        EmployeeService.updateEmployee(userId, employee, employee.uuid).then(
+        var records = [];
+        for (var i = 0; i < $scope.customizedEmployeeFields.length; i++) {
+            var record = new Record($scope.customizedEmployeeFields[i].id, employee.uuid, LoginService.getUserInfo().userId, $scope.customizedEmployeeFields[i].value);
+            records.push(record);
+        }
+        employee.records = records;
+        $scope.employeeEditPromise = EmployeeService.updateEmployee(userId, employee, employee.uuid).then(
             function (data) {
                 var list = MessageService.handleMsg(data);
                 if (list) {
-                    $scope.editedEmployee={uuid:-1};
+                    $scope.editedEmployee = {uuid: -1};
                     $scope.employees = list;
-                    newEmployee = {isNew:true};
+                    newEmployee = {isNew: true};
                     $scope.employees.push(newEmployee);
                     $scope.editing = false;
                 }
@@ -131,7 +139,199 @@ app.controller('UserEmployeeController', ['$scope', 'EmployeeService', 'LoginSer
         );
     };
 
+    $scope.customizedEmployeeFields = [];
+    var getCustomizedField = function () {
+        $scope.employeeAllPromise = EmployeeService.getEmployeeCustomizedField(LoginService.getUserInfo().userId)
+            .then(
+            function (data) {
+                var f = MessageService.handleMsg(data);
+                if (f) {
+                    $scope.customizedEmployeeFields = f;
+                }
+            },
+            function (err) {
+                MessageService.handleServerErr(err);
+            }
+        );
+    };
+
+    //var getCustomizedEmployeeFieldsValue  = function (employee) {
+    //    setCustomizedEmployeeFieldsValue($scope.customizedEmployeeFields,employee.records);
+    //}
+
+    var setCustomizedEmployeeFieldsValue = function (fields, values) {
+        for (var i = 0; i < fields.length; i++) {
+            var j = 0;
+            for (j = 0; j < values.length; j++) {
+                if (fields[i].id == values[j].questionId) {
+                    fields[i].value = values[j].value;
+                    break;
+                }
+            }
+            if (j == values.length) {
+                fields[i].value = '';
+            }
+        }
+    }
+    var Record = function (questionId, oe_id, user_id, value) {
+        this.questionId = questionId;
+        this.oeId = oe_id;
+        this.userId = user_id;
+        this.value = value;
+    };
+    //var saveCustomizedEmployeeFieldsValue = function (employeeId) {
+    //    var records = [];
+    //    for(var i=0;i<$scope.customizedEmployeeFields.length;i++){
+    //        var record = new Record($scope.customizedEmployeeFields[i].id,employeeId,LoginService.getUserInfo().userId,$scope.customizedEmployeeFields.value);
+    //        records.push(record);
+    //    }
+    //
+    //    $scope.employeeEditPromise = UserFormRecordService.saveQuestionValues(records)
+    //        .then(
+    //        function (data) {
+    //            MessageService.handleMsg(data);
+    //        },
+    //        function (err) {
+    //            MessageService.handleServerErr(err);
+    //        }
+    //    );
+    //}
+
+    // view report
+
+    $scope.selectAndShowEmployee = function (employee) {
+        angular.forEach($scope.employees, function (rep) {
+            rep.selected = rep.uuid == employee.uuid;
+        });
+        $scope.viewEmployeeReportPromise = EmployeeService.getEmployee(LoginService.getUserInfo().userId, $scope.employees[index].uuid)
+            .then(
+            function (msg) {
+                var employee = MessageService.handleMsg(msg);
+                if (employee) {
+                    $timeout(function () {
+                        var reports=[];
+                        reports.push($scope.folderTree.currentNode);
+                        handleEmployeeReports(reports, employee);
+                    });
+                }
+            },
+            function (err) {
+                MessageService.handleServerErr(err);
+            }
+        );
+
+
+    };
+
+    $scope.employeeReports = [];
+    $scope.viewReports = function (employee) {
+        if ($scope.employeeReports.length == 0) {
+            $scope.employeeEditPromise = UserFolderService.getUserEmployeeReports(LoginService.getUserInfo().userId)
+                .then(
+                function (response) {
+                    var rep = MessageService.handleMsg(response);
+                    if (rep) {
+                        $scope.employeeReports = rep;
+                        handleEmployeeReports($scope.employeeReports, employee);
+                        $scope.editedEmployee.viewingReports = true;
+                    }
+                },
+                function (err) {
+                    MessageService.handleServerErr(err);
+                }
+            );
+        } else {
+            handleEmployeeReports($scope.employeeReports, employee);
+            $scope.editedEmployee.viewingReports = true;
+        }
+    };
+
+    var handleEmployeeReports = function (reports, employee) {
+        var customizedFields = employee.records;
+        for (var i = 0; i < reports.length; i++) {
+            var report = reports[i];
+            var cfs = report.questions;
+            var finish = 0;
+            for (var k = 0; k < cfs.length; k++) {
+                var j = 0;
+                for (j = 0; j < customizedFields.length; j++) {
+                    if (customizedFields[j].questionId == cfs[k].id) {
+                        cfs[k].value = customizedFields[j].value;
+                        if (cfs[k].value && cfs[k].value != '') {
+                            finish++;
+                        }
+                        break;
+                    }
+                }
+                if (j == customizedFields.length || cfs[k].value == '') {
+                    cfs[k].value = '{PLEASE FILL FIELD ' + cfs[k].name + ' }';
+                }
+            }
+            report.percent = 100 * finish / cfs.length;
+        }
+        $scope.selectAndShowReport(reports[0]);
+    };
+
+    $scope.selectAndShowReport = function (report) {
+        angular.forEach($scope.employeeReports, function (rep) {
+            rep.selected = rep.id == report.id;
+        });
+        $timeout(function () {
+            viewPDF(report);
+        });
+
+    };
+
+    var viewPDF = function (file) {
+        var formElement = angular.element(file.content.replace(/{-/gi, '').replace(/-}/gi, ''));
+        var records = file.questions;
+        for (var j = 0; j < records.length; j++) {
+            var record = records[j];
+            var record_question_id = record.id;
+            var questionElement = angular.element(formElement.find("plugin[question_id='" + record_question_id + "']"));
+            var span = angular.element("<span></span>");
+            span.text(record.value);
+            questionElement.replaceWith(span);
+        }
+        angular.element('#form_container').html('').append(formElement);
+        $scope.showPDF();
+    }
+    // show pdf
+    $scope.showPDF = function () {
+        var pdf = new jsPDF('p', 'pt', 'a4');
+        var source = $('#form_container').get(0);
+        angular.element(source).css("display", "block");
+        var specialElementHandlers = {
+            '#bypassme': function (element, renderer) {
+                return true
+            }
+        };
+        var margins = {
+            top: 80,
+            bottom: 60,
+            left: 40,
+            width: 522
+        };
+        // all coords and widths are in jsPDF instance's declared units
+        // 'inches' in this case
+        pdf.fromHTML(
+            source, // HTML string or DOM elem ref.
+            margins.left, // x coord
+            margins.top, { // y coord
+                'width': margins.width, // max width of content on PDF
+                'elementHandlers': specialElementHandlers
+            },
+
+            function (dispose) {
+                $scope.content = pdf.output("dataurlstring");
+                //pdf.autoPrint();  // <<--------------------- !!
+                //pdf.output('dataurlnewwindow');
+            }, margins);
+        angular.element(source).css("display", "none");
+    };
+
 }]);
+
 
 //
 //app.controller('UserEmployeeEditController', ['$scope', 'EmployeeService', 'UserEmploymentStatusService', 'UserJobTitleService', '$location', 'FileUploader', 'LoginService', 'usSpinnerService', function ($scope, EmployeeService, UserEmploymentStatusService, UserJobTitleService, $location, FileUploader, LoginService, usSpinnerService) {
